server/flask-server/requirements.txt
flask
python-dotenv
psycopg2-binary
pyjwt
bcrypt
flask-cors
requests
python-dotenv